Move coalesceLatest operator into Sources/CoalesceLatestPublisher.swift
cmux policy: major added types get their own TypeName.swift file. Wires
the new file into the pbxproj alongside WorkspaceSidebarObservation.swift.
